+++ linter/linter.go
@@ -32,6 +32,11 @@ func Lint(r io.Reader) (*validateachangelog.Changelog, error) {
 	for scanner.Scan() {
 		line := scanner.Text()
 
+		// Parse title
+		if internal.IsTitleLine(line) {
+			c.Title = internal.ParseTitleLine(line)
+		}
+
 		// Parse version
 		if strings.HasPrefix(line, "## ") {
 			// Push current version if there is one