+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Introduce cmd/lint-changelog.
+- Parse changelog title.
 
 ## [0.3.0] - 2025-10-28